requirements.txt
fastapi
uvicorn
earthengine-api
python-dotenv
